moved dialog specific axesc andctypes to nlp/neural_types.py, refactored the modules
