+++ nemo/collections/nlp/nm/non_trainables/dialogue_state_tracking/system_utterance_history_update.py
@@ -21,6 +21,7 @@ https://github.com/thu-coai/ConvLab-2/
 
 from nemo.backends.pytorch.nm import NonTrainableNM
 from nemo.core.neural_types import *
+from nemo.collections.nlp.neural_types import *
 from nemo.utils import logging
 from nemo.utils.decorators import add_port_docs
