+++ nemo/collections/nlp/nm/non_trainables/dialogue_state_tracking/user_utterance_encoder.py
@@ -22,6 +22,7 @@ import torch
 
 from nemo.backends.pytorch.nm import NonTrainableNM
 from nemo.core.neural_types import *
+from nemo.collections.nlp.neural_types import *
 from nemo.utils import logging
 from nemo.utils.decorators import add_port_docs
