+++ nemo/core/neural_types/elements.py
@@ -43,14 +43,8 @@ __all__ = [
     'NormalizedImageValue',
     'StringLabel',
     'StringType',
-    'Utterance',
-    'UserUtterance',
-    'SystemUtterance',
-    'AgentUtterance',
     'TokenIndex',
     'Length',
-    'SlotValue',
-    'MultiWOZDomainState',
 ]
 
 import abc
@@ -256,28 +250,6 @@ class StringLabel(StringType):
     """
 
 
-class Utterance(StringType):
-    """Element type representing an utterance (e.g. "Is there a train from Ely to Cambridge on Tuesday ?")."""
-
-
-class UserUtterance(Utterance):
-    """Element type representing a utterance expresesd by the user."""
-
-
-class SystemUtterance(Utterance):
-    """Element type representing an utterance produced by the system."""
-
-
-class AgentUtterance(ElementType):
-    """Element type representing utterance returned by an agent (user or system) participating in a dialog."""
-
-    # def __str__(self):
-    #    return "Utterance returned by an agent (user or system) participating in a dialog."
-
-    # def fields(self):
-    #    return ("agent", "utterance")
-
-
 class IntType(ElementType):
     """Element type representing a single integer"""
 
@@ -288,20 +260,3 @@ class TokenIndex(IntType):
 
 class Length(IntType):
     """Type representing an element storing a "length" (e.g. length of a list)."""
-
-
-class SlotValue(ElementType):
-    """Element type representing slot-value pair."""
-
-    # def __str__(self):
-    #    return "Slot-value pair"
-
-    # def fields(self):
-    #    return ("slot", "value")
-
-
-class MultiWOZDomainState(SlotValue):
-    """Element type representing MultiWOZ slot-value pair."""
-
-    # def fields(self):
-    #    return ("book", "semi")
